fix erro de rename de pacotes
